pyproject.toml
[tool.poetry]
name = "mlsat-satnogs"
version = "0.1.0"
description = ""
authors = ["sakhn <sakhnikarseni@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
python-dotenv = "^1.0.0"
pymongo = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
